+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Option for quiet level.
 - Enforce prefixes in label names.
 - Warn about +-.
+- Enforce hard space before `\ref`.
 
 ### Fixed